Major infrastructure changes to newtestsuite include

+++ newtestsuite/src/runTests.C
@@ -1,4 +1,4 @@
-#ifdef os_windows
+#ifdef os_windows_test
 //needed for Sleep
 #include <windows.h>
 #define sleep(x) Sleep(x * 1000)
@@ -17,6 +17,7 @@
 
 #include "runTests-utils.h"
 #include "error.h"
+#include "help.h"
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <cstring>
@@ -25,9 +26,13 @@
 #define MAX_ITER 1000
 
 // Default name for the resume log file
-#define DEFAULT_RESUMELOG "resumelog"
+#define DEFAULT_RESUMELOG "bresumelog"
 // Default name for the crash log file
 #define DEFAULT_CRASHLOG "crashlog"
+
+#define MEMCPU_DEFAULT_LOG "memcpu_tmp.log";
+const char *memcpu_name = NULL;
+const char *memcpu_orig_name = NULL;
 
 bool staticTests = false;
 bool useLog = false;
@@ -38,6 +43,46 @@ string pdscrdir;
 int testLimit = 10;
 
 vector<char *> child_argv;
+
+void parseMEMCPUFile()
+{
+   if (!memcpu_name || !memcpu_orig_name)
+      return;
+
+   signed long mem_total = 0, utime_total = 0, stime_total = 0;
+   FILE *f = fopen(memcpu_name, "r");
+   if (!f)
+      return;
+
+   for (;;)
+   {
+      signed long mem, utime, stime;
+      int res = fscanf(f, "mem=%ld\tutime=%ld\tstime=%ld\n",
+                       &mem, &utime, &stime);
+      if (res != 3)
+         break;
+      mem_total += mem;
+      utime_total += utime;
+      stime_total += stime;
+   }
+   fclose(f);
+   unlink(memcpu_name);
+
+   if (strcmp(memcpu_orig_name, "-") == 0)
+   {
+      f = stdout;
+   }
+   else {
+      f = fopen(memcpu_orig_name, "w");
+      if (!f)
+         return;
+   }
+   
+   fprintf(f, "mem=%ld\tutime=%ld\tstime=%ld\n",
+           mem_total, utime_total, stime_total);
+   if (f != stdout)
+      fclose(f);
+}
 
 // isRegFile:
 // Returns true if filename is a regular file
@@ -90,41 +135,44 @@ void getInput(const char *filename, string& output)
 void parseParameters(int argc, char *argv[])
 {
 #if defined(STATIC_TEST_DRIVER)
-  staticTests = true;
+   staticTests = true;
 #endif
 
    for ( int i = 1; i < argc; i++ )
    {
-      if ( strncmp(argv[i], "-log", 4) == 0 )
-      {
-         useLog = true;
-         // Check to see if next line provides a logfile name
-         if ((i + 1 < argc)
-	     && ((argv[i+1][0] != '-' ) || (strcmp(argv[i+1], "-") == 0)))
-         {
-            logfile = argv[++i];
+      if (strncmp(argv[i], "-limit", 6) == 0) {
+         if (i == (argc-1)) {
+            fprintf(stderr, "Error: -limit requires a parameter\n");
          }
-      }
-      else if (strncmp(argv[i], "-limit", 6) == 0) {
-          if (i == (argc-1)) {
-              fprintf(stderr, "Error: -limit requires a parameter\n");
-          }
-          else {
-              unsigned int limit = atoi(argv[i+1]);
-              testLimit = limit;
-              i++;
-          }
-      } else if (strcmp(argv[i], "-help") == 0) {
-	printf("Usage: runTests [-log [file]] [-limit <limit>] [-static | -dynamic] ...\n");
-	printf("\t-log: enable logging, and use the file specified if one is provided\n");
-	printf("\t\t(to log to standard output, use '-log -')\n");
-	printf("runTests also passes parameters to test_driver:\n");
-	system("test_driver -help");
-	exit(0);
+         else {
+            unsigned int limit = atoi(argv[i+1]);
+            testLimit = limit;
+            i++;
+         }
+      } else if ((strcmp(argv[i], "-help") == 0) ||
+                 (strcmp(argv[i], "--help") == 0)) {
+         print_help();
+         exit(0);
       } else if (strcmp(argv[i], "-static") == 0) {
-	staticTests = true;
+         staticTests = true;
       } else if (strcmp(argv[i], "-dynamic") == 0) {
-	staticTests = false;
+         staticTests = false;
+      }
+      else if ((strcmp(argv[i], "-memcpu") == 0) ||
+               (strcmp(argv[i], "-cpumem") == 0))
+      {
+         memcpu_name = MEMCPU_DEFAULT_LOG;
+         
+         if ((i+1 < argc) &&
+             (argv[i+1][0] != '-' || argv[i+1][1] == '\0'))
+         {
+            i++;
+            memcpu_orig_name = argv[i];
+         }
+         else
+         {
+            memcpu_orig_name = "-";
+         }
       }
       else
       {
@@ -169,47 +217,55 @@ int main(int argc, char *argv[])
    // Remove a stale resumelog, if it exists
    if ( getenv("RESUMELOG") && isRegFile(string(getenv("RESUMELOG"))) )
    {
-     unlink(getenv("RESUMELOG"));
+      unlink(getenv("RESUMELOG"));
    } else if (isRegFile(string(DEFAULT_RESUMELOG))) {
-     unlink(DEFAULT_RESUMELOG);
+      unlink(DEFAULT_RESUMELOG);
    }
 
    // Remove a stale crashlog, if it exists
    if (getenv("CRASHLOG") && isRegFile(string(getenv("CRASHLOG")))) {
-     unlink(getenv("CRASHLOG"));
+      unlink(getenv("CRASHLOG"));
    } else if (isRegFile(string(DEFAULT_CRASHLOG))) {
-     unlink(DEFAULT_CRASHLOG);
+      unlink(DEFAULT_CRASHLOG);
    }
 
    // Create a PIDs file, to track mutatee PIDs
-   char pidFilename[32];
-   initPIDFilename(pidFilename, 32);
-   if (isRegFile(string(pidFilename))) {
-     unlink(pidFilename); // Ensure that the file doesn't already exist
-   }
+   char *pidFilename = NULL;
 
    // result == 2 indicates that there are no more tests to run
    while ( result != NOTESTS && invocation < MAX_ITER )
    {
       result = RunTest(invocation, useLog, staticTests, logfile, testLimit,
-		       child_argv, pidFilename);
+                       child_argv, pidFilename, memcpu_name);
       invocation++;
       // I want to kill any remaining mutatees now, to clean up.  I should also
       // set a timer in RunTest in case something goes weird with test_driver.
       // (I think we'd be better off moving away from timer.pl)
       cleanupMutatees(pidFilename);
       if (-3 == result) {
-	// User interrupted the test run; allow them a couple of seconds to do
-	// it again and kill runTests
-	// TODO Make sure this is portable to Windows
-	sleep(2);
+         // User interrupted the test run; allow them a couple of seconds to do
+         // it again and kill runTests
+         // TODO Make sure this is portable to Windows
+         fprintf(stderr, "Press ctrl-c again with-in 2 seconds to abort runTests.\n");
+         sleep(2);
+      }
+      if (-4 == (signed char) result) {
+         fprintf(stderr, "Could not execute test_driver\n");
+         break;
+      }
+      if (-5 == (signed char) result) {
+         break;
       }
    }
 
    // Remove the PID file, now that we're done with it
-   if (isRegFile(string(pidFilename))) {
-     unlink(pidFilename);
+   if (pidFilename && isRegFile(string(pidFilename))) {
+      unlink(pidFilename);
    }
+   unlink(DEFAULT_RESUMELOG);
+   unlink("");
+   
 
+   parseMEMCPUFile();
    return 0;
 }
